Implement WhatsApp bulk messaging automation with media support and GUI integration
