--- numbers.txt
@@ -1,5 +1,0 @@
-9005434580524
-9005307677806
-9005307677806
-9005307677806
-9005307677806
