+++ old/automator2.py
